Move recipe for Renesas Graphic Feature to another layer
Renesas Graphic Feature will be provided by RZ feature layer.
RZ feature layer is named 'meta-rz-feature'.
If none of that layer, we will use CPU for Graphic feature instead.

We also can enable/disable using Renesas Graphic Feature by
USE_RENESAS_GLES option in local.conf

+++ recipes-graphics/wayland/wayland_%d.bbappend
@@ -1,11 +1,7 @@
-do_install_append_rzg2l() {
-	# These files is provided in other recipes
-	rm -rf ${D}/${libdir}/libwayland-egl*
-	rm -rf ${D}/${libdir}/pkgconfig/wayland-egl*
-}
-
-do_install_append_rzv2l() {
-       # These files is provided in other recipes
-       rm -rf ${D}/${libdir}/libwayland-egl*
-       rm -rf ${D}/${libdir}/pkgconfig/wayland-egl*
+do_install_append() {
+	if [ "${EXT_GFX_BACKEND}" = "1" ]; then
+		# These files are provided in other recipes
+		rm -rf ${D}/${libdir}/libwayland-egl*
+		rm -rf ${D}/${libdir}/pkgconfig/wayland-egl*
+	fi
 }
